cloud_function/requirements.txt
functions-framework==3.*
requests>=2.28.0
google-cloud-secret-manager>=2.16.0
